remove CaptureScreenController and add join(), close(), cancel()
