--- ui/components/capture/src/main/java/com/google/jetpackcamera/ui/components/capture/controller/CaptureScreenController.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2026 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.jetpackcamera.ui.components.capture.controller
-
-import com.google.jetpackcamera.model.DeviceRotation
-
-/**
- * Interface for controlling actions on the capture screen.
- */
-interface CaptureScreenController {
-    /**
-     * Sets the display rotation.
-     *
-     * @param deviceRotation The device rotation to set.
-     */
-    fun setDisplayRotation(deviceRotation: DeviceRotation)
-
-    /**
-     * Updates the UI with the last captured media.
-     */
-    fun updateLastCapturedMedia()
-}
