+++ ui/components/capture/src/main/java/com/google/jetpackcamera/ui/components/capture/controller/ImageWellController.kt
@@ -27,4 +27,9 @@ interface ImageWellController {
      * @param mediaDescriptor The media descriptor to be transferred.
      */
     fun imageWellToRepository(mediaDescriptor: MediaDescriptor)
+
+    /**
+     * Updates the UI with the last captured media.
+     */
+    fun updateLastCapturedMedia()
 }
