+++ ui/components/capture/src/main/java/com/google/jetpackcamera/ui/components/capture/controller/ImageWellControllerImpl.kt
@@ -27,10 +27,12 @@ import kotlinx.coroutines.launch
  * Implementation of [ImageWellController] that handles image well actions.
  *
  * @param mediaRepository The [MediaRepository] for accessing media.
+ * @param updateLastCapturedMediaCallback Callback to update the last captured media.
  * @param coroutineContext The [CoroutineContext] for launching coroutines.
  */
 class ImageWellControllerImpl(
     private val mediaRepository: MediaRepository,
+    private val updateLastCapturedMediaCallback: () -> Unit,
     coroutineContext: CoroutineContext
 ) : ImageWellController {
     private val job = Job(parent = coroutineContext[Job])
@@ -42,5 +44,20 @@ class ImageWellControllerImpl(
                 mediaDescriptor
             )
         }
+    }
+    override fun updateLastCapturedMedia() {
+        updateLastCapturedMediaCallback()
+    }
+
+    suspend fun join() {
+        job.join()
+    }
+
+    fun cancel() {
+        job.cancel()
+    }
+
+    fun close() {
+        job.cancel()
     }
 }
